Implement UI element distance sorting
• UI elements are now sorted by distance from playeras well as zOffset. Their HudSpace origins are used to calculate distance from the player.
• Rough implementation, will require debugging, benchmarking and optimization before being pushed to main.
• Max distance ~8191 meters
• Min precision: ~12.5 cm per unit

Typically, semitransparent geometry is sorted based on distance to the camera, but the billboards supplied by the mod API aren't sorted at all. They're drawn in the order added. I'd like to think this was done intentionally to give modders the option of customizing sorting or omitting it altogether, but I can't help but think it was just a mistake.