Optimize update accessor sorting and reduce sorting range
• Distance now only calculated for each unique node position delegate
• Sorting now done at 1/10th normal speed
• zOffset and distance sorting now done simultaneously instead of separately
• Max sorting distance now clamped to 655.35m with a precsion of 1cm/unit

Given that the vast majority of UI elements will be relying only on a handful of HudSpaceNodes, the number of unique position delegates supplied will be very small. There is no need to recalculate the distance for every UI element, as the distance reported rarely changes. Taken together with SE's absurd profiler overhead, it quickly becomes preferable to weed out the unique delegates with a HashSet and build a dictionary with the distances precalculated for each delegate instance.