Replace camera freezing logic by cancelling rotation of controlled ent
I've tried a bunch of different ways to prevent mouse movement from moving the camera, usually involving spectator camera and manipulating the world matrix of the ply char entity. 

I've since found that simply calling IMyControllableEntity.MoveAndRotate() and passing in the last motion indicator (sans rotation vec), I can achieve the same effect, but without any of the apparent limitations of the previous solutions. It seems to be functionally perfect.