Refine Bind Manager input filter to identify potential new input sequence
Rather than base input filtering on the total number of controls pressed, the Bind manager will now attempt to identify potential bind sequences based on when these controls are pressed. To be considered part of a potential sequence, a control must be newly pressed and for multiple controls to be associated with eachother, they must be pressed within one second of the last one added to the sequence.

This length, rather than the content, of the resulting sequence is then used to filter input for key binds. If the candidate sequence is longer than the number of controls pressed within a bind group, then the bind groups affected will not invoke any binds. The pressed control count within a bind group is determined by the number of controls in otherwise pressed, disambiguated and unique key binds.

In theory this should be a fair bit more forgiving than the last solution, while still accomplishing the primary purpose of excluding most input that isn't meant for a given bind group.