Use two buffers for BB rendering
In the previous iteration, updating these billboards during Draw would cause displacement problems, where on some frames, the billboards would lag behind camera movement. It seems to be some sort of timing/sync problem. This behavior is similar to the sorts of issues produced by attempting to update from the wrong update method, or from another thread.

Assuming this is caused by some sort of thread synchronization problem, I've opted to use a pair of alternating front/back buffer lists for billboards. One to use while I'm drawing and another for the render to use while it's doing whatever it's doing. This solution appears to work, but I'm not sure how the renderer is supposed to handle synchronization, so it still might not be correct.