Fix bug preventing BindGroups from releasing partly pressed binds
Previously the beingReleased flag was only updated if at least one bind was pressed. If a bind were partly pressed for one frame, then on the next frame, no binds in that group were pressed, this flag would stop being evaluated.

Consequently, the next time a bind in that group is pressed that bind that had been partially pressed, and if the newly pressed bind were shorter but overlapped with the last bind, previously would still  appear to be partly pressed as if no time had passed. Creating a bind conflict where none should exist.