Fix race condition in TreeManager
The UI tree is only partially updated once(ish) per frame, on a separate thread. This presents a problem when the client is attempting to unregister itself from the API, as it's possible Master will hold onto delegates to client-side objects in an undefined state and invoke them.

I've corrected the problem by forcing the tree to perform a full update if one or more clients were unregistered since the last frame.