Fix race condition in TreeManager
• Fixed race condition caused by missing layout buffer.

UI tree updates are typically spread out over at least three ticks, one per frame, on a separate thread. Race conditions are avoided by having separate buffers for the update delegate lists, only I forgot to add a buffer for one of the lists. This error manifested so rarely because tree updates only take a fraction of a ms, and occur at the end of the frame.

• Tree updates now start/finish at the beginning of the frame.